refactor(cli): extract explore feishu sink commands into their own module
explore.py was the largest cli_commands module at 989 lines, only 11 under
the modularization smoke's default budget, as the remaining oversized seam
after the quota/status/todo/history/scheduler plumbing extractions. Move
the cohesive feishu-* sink family (feishu-setup, feishu-visual-configure,
feishu-sync, feishu-card) -- registration, dispatch branches, and the
sync-only _target_config helper -- into
loopx/cli_commands/explore_feishu_commands.py, mirroring the
explore_planning_commands sibling pattern with shared arg helpers injected
as callables. explore.py drops to 723 lines.

Public invocations are unchanged: loopx.cli still calls
register_explore_commands/handle_explore_command, subcommand flags and
explore --help output are byte-identical, and the singleflight smoke now
patches the moved names in their new namespace instead of keeping a
compatibility alias.

Also drop the stale LEGACY_MODULE_LIMITS entries for
benchmark_review_lifecycle.py and terminal_bench_environment_result.py:
the native-runner benchmark reset (#3267) removed both modules, which left
the size/ownership smoke failing on main with "size budgets reference
missing modules".
